add back Context type as alias to context.Context
Pull request #131 deleted the Context declaration,
which broke most usages of this package since users needed to implement
the interfaces mentioned in this package by referencing the
groupcache.Context type. However, that type was removed from peers.go
in the mentioned pull request.

Technically, #131 is a breaking change since groupcache.Context
was an interface{}, and the PR switched it to be context.Context.
Fortunately, it seems that all users are passing a context.Context
anyways so that the type safety just checks out.

However, most users are still referencing groupcache.Context.